[Config]: Proper defines for framerate variying support. (60hz even tested, works lol.) [Client]: Some lacking int64_t support in demo.cpp [Net]: Added typecasting accomodating to config.h changes. [Server]: Patched SV_GiveMSec(had a hardcoded 16 integral).

+++ src/common/net/net.cpp
@@ -567,7 +567,7 @@ static bool NET_SendLoopPacket(netsrc_t sock, const void *data,
     loopback_t *loop;
     loopmsg_t *msg;
 
-    if (net_dropsim->integer > 0 && (Q_rand() % BASE_FRAMETIME) < net_dropsim->integer) {
+    if (net_dropsim->integer > 0 && (Q_rand() % (int64_t)BASE_FRAMETIME) < net_dropsim->integer) {
         return false;
     }
